+++ cmd/snap/main.go
@@ -24,7 +24,6 @@ import (
 	"io"
 	"os"
 	"path/filepath"
-	"runtime"
 	"strings"
 	"unicode"
 	"unicode/utf8"
@@ -295,17 +294,6 @@ var ClientConfig = client.Config{
 	Socket: dirs.SnapdSocket,
 	// Allow interactivity if we have a terminal
 	Interactive: isStdinTTY,
-}
-
-// Client returns a new client using ClientConfig as configuration.
-func Client() *client.Client {
-	if runtime.GOOS != "linux" {
-		fmt.Fprintf(Stderr, i18n.G(`Interacting with snapd is not yet supported on %s.
-This command has been left available for documentation purposes only.
-`), runtime.GOOS)
-		os.Exit(1)
-	}
-	return client.New(&ClientConfig)
 }
 
 func init() {
